model_server/pyproject.toml
[tool.poetry]
name = "model-server"
version = "0.1.0"
description = ""
authors = ["ddymkowski <damiandymkowski96@gmail.com>"]
readme = "README.md"
packages = [{include = "model_server"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.95.0"}
torch = "^2.0.0"
pillow = "^9.4.0"
torchvision = "^0.15.1"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
mypy = "^1.1.1"
isort = "^5.12.0"
types-pillow = "^9.4.0.17"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



[tool.mypy]
ignore_missing_imports = true